requirements.txt
future~=0.18.3
pillow~=10.2.0
opencv-python~=4.9.0.80
numpy~=1.26.4
tensorflow~=2.16.1
matplotlib~=3.5.2
scikit-learn~=1.4.1.post1